Jordan,
Started laying out the basic structure for us. This way you can jump right into playing with the data/visualizations aspect. I'll get some more things put into this tomorrow night and/or Tuesday night after you take a crack at it. In particular, I'll try to get a working version of that saving to HTML functionality included since the client seems to be in big favor of that piece.

Dom